requirements.txt
# Requirements for: XDenseQNet -- Hybrid Quantum Neural Network for Skin Lesion Classification
# Python 3.8+ required
# Install with: pip install -r requirements.txt
# For GPU support: install CUDA 11.8+ and the GPU version of torch

# Core ML
torch>=2.1.0
torchvision>=0.16.0

# Quantum ML
pennylane>=0.36.0
pennylane-lightning>=0.36.0

# Pre-trained models (ConvNeXt, Swin Transformer)
timm>=1.0.0

# Data augmentation
albumentations>=1.3.0

# Data & Metrics
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
PyYAML>=6.0.1

# Utilities
tqdm>=4.66.0

# Development / Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
pytest>=7.4.0
